MongoDB Sucessfully Connected
Took some time but got the database connection working. Turns out I really didn't have to  change all that much just add the documents in the database and name the collections and database correctly in the files. The rest from the tutorail was enough. I got to see many errors on the way, too much to list down all but most of them were due easy to find and some of them don't bother because if the connection and files are right you won't come across them